XP-1917 Move PortalHandler (and friends) to portal-api
PortalException -> com.enonic.xp.portal.PortalException
PortalHandler -> com.enonic.xp.portal.handler.PortalHandler
BaseHandler -> com.enonic.xp.portal.handler.BaseHandler
PortalHandlerWorker -> com.enonic.xp.portal.handler.PortalHandlerWorker
EndpointHandler -> com.enonic.xp.portal.handler.EndpointHandler
ControllerScript -> com.enonic.xp.portal.controller.ControllerScript
ControllerScriptFactory -> com.enonic.xp.portal.controller.ControllerScriptFactory
