+++ modules/portal/portal-api/src/main/java/com/enonic/xp/portal/controller/ControllerScript.java
@@ -1,9 +1,9 @@
-package com.enonic.xp.portal.impl.controller;
+package com.enonic.xp.portal.controller;
 
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
 
 public interface ControllerScript
 {
-    public PortalResponse execute( PortalRequest portalRequest );
+    PortalResponse execute( PortalRequest portalRequest );
 }
