+++ modules/portal/portal-api/src/main/java/com/enonic/xp/portal/handler/PortalHandler.java
@@ -1,4 +1,4 @@
-package com.enonic.xp.portal.impl;
+package com.enonic.xp.portal.handler;
 
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
