+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/PortalServlet.java
@@ -25,8 +25,10 @@ import com.enonic.xp.branch.Branch;
 import com.enonic.xp.content.ContentPath;
 import com.enonic.xp.context.ContextAccessor;
 import com.enonic.xp.portal.PortalAttributes;
+import com.enonic.xp.portal.PortalException;
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
+import com.enonic.xp.portal.handler.PortalHandler;
 import com.enonic.xp.portal.impl.exception.ExceptionMapper;
 import com.enonic.xp.portal.impl.exception.ExceptionRenderer;
 import com.enonic.xp.portal.impl.serializer.RequestBodyReader;
