+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/controller/ControllerScriptFactoryImpl.java
@@ -3,6 +3,8 @@ package com.enonic.xp.portal.impl.controller;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
+import com.enonic.xp.portal.controller.ControllerScript;
+import com.enonic.xp.portal.controller.ControllerScriptFactory;
 import com.enonic.xp.portal.script.PortalScriptService;
 import com.enonic.xp.resource.ResourceKey;
 import com.enonic.xp.script.ScriptExports;
