+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/exception/ExceptionMapper.java
@@ -1,8 +1,8 @@
 package com.enonic.xp.portal.impl.exception;
 
 import com.enonic.xp.exception.NotFoundException;
+import com.enonic.xp.portal.PortalException;
 import com.enonic.xp.portal.PortalResponse;
-import com.enonic.xp.portal.impl.PortalException;
 import com.enonic.xp.web.HttpStatus;
 
 public final class ExceptionMapper
