+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/exception/ExceptionRendererImpl.java
@@ -5,9 +5,9 @@ import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.enonic.xp.portal.PortalException;
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
-import com.enonic.xp.portal.impl.PortalException;
 import com.enonic.xp.resource.ResourceService;
 
 @Component
