+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/handler/attachment/AttachmentHandlerWorker.java
@@ -8,7 +8,7 @@ import com.enonic.xp.attachment.Attachments;
 import com.enonic.xp.content.Content;
 import com.enonic.xp.content.ContentId;
 import com.enonic.xp.content.ContentService;
-import com.enonic.xp.portal.impl.handler.PortalHandlerWorker;
+import com.enonic.xp.portal.handler.PortalHandlerWorker;
 import com.enonic.xp.web.HttpStatus;
 
 final class AttachmentHandlerWorker
