+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/handler/controller/ControllerHandlerWorker.java
@@ -6,7 +6,7 @@ import com.enonic.xp.content.ContentNotFoundException;
 import com.enonic.xp.content.ContentPath;
 import com.enonic.xp.content.ContentService;
 import com.enonic.xp.portal.RenderMode;
-import com.enonic.xp.portal.impl.handler.PortalHandlerWorker;
+import com.enonic.xp.portal.handler.PortalHandlerWorker;
 import com.enonic.xp.site.Site;
 
 public abstract class ControllerHandlerWorker
