+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/handler/image/ImageHandlerWorker.java
@@ -15,7 +15,7 @@ import com.enonic.xp.content.Media;
 import com.enonic.xp.image.ImageService;
 import com.enonic.xp.image.ReadImageParams;
 import com.enonic.xp.image.ScaleParams;
-import com.enonic.xp.portal.impl.handler.PortalHandlerWorker;
+import com.enonic.xp.portal.handler.PortalHandlerWorker;
 import com.enonic.xp.util.MediaTypes;
 import com.enonic.xp.web.HttpStatus;
