+++ modules/portal/portal-impl/src/main/java/com/enonic/xp/portal/impl/rendering/DescriptorBasedComponentRenderer.java
@@ -12,8 +12,8 @@ import com.enonic.xp.page.DescriptorKey;
 import com.enonic.xp.portal.PortalRequest;
 import com.enonic.xp.portal.PortalResponse;
 import com.enonic.xp.portal.RenderMode;
-import com.enonic.xp.portal.impl.controller.ControllerScript;
-import com.enonic.xp.portal.impl.controller.ControllerScriptFactory;
+import com.enonic.xp.portal.controller.ControllerScript;
+import com.enonic.xp.portal.controller.ControllerScriptFactory;
 import com.enonic.xp.portal.rendering.Renderer;
 import com.enonic.xp.region.Component;
 import com.enonic.xp.region.Descriptor;
