+++ modules/portal/portal-impl/src/test/java/com/enonic/xp/portal/impl/exception/ExceptionMapperTest.java
@@ -4,8 +4,8 @@ import org.junit.Before;
 import org.junit.Test;
 
 import com.enonic.xp.exception.NotFoundException;
+import com.enonic.xp.portal.PortalException;
 import com.enonic.xp.portal.PortalResponse;
-import com.enonic.xp.portal.impl.PortalException;
 import com.enonic.xp.web.HttpStatus;
 
 import static org.junit.Assert.*;
