+++ modules/portal/portal-impl/src/test/java/com/enonic/xp/portal/impl/handler/error/ErrorHandlerTest.java
@@ -2,9 +2,9 @@ package com.enonic.xp.portal.impl.handler.error;
 
 import org.junit.Test;
 
+import com.enonic.xp.portal.PortalException;
 import com.enonic.xp.portal.PortalResponse;
-import com.enonic.xp.portal.impl.PortalException;
-import com.enonic.xp.portal.impl.handler.BaseHandlerTest;
+import com.enonic.xp.portal.handler.BaseHandlerTest;
 import com.enonic.xp.web.HttpMethod;
 import com.enonic.xp.web.HttpStatus;
